+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Increase the lookback interval for weather data to 1.5 hours in alerts to prevent problems with the first data point.
 * Report Overwatcher alerts in the `/alerts` endpoint.
 
+
 ## 0.2.11 - 2025-04-01
 
 ### 🚀 New